backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
onnxruntime
Pillow
numpy
